Server/Python/python_socketio/requirements.txt
python-socketio[client]==5.6.0
numpy==1.23.5